+++ sdk/go/consts.go
@@ -7,8 +7,8 @@ import "time"
 const (
     DatastarKey = "datastar"
     Version                   = "1.0.0-beta.9"
-    VersionClientByteSize     = 36027
-    VersionClientByteSizeGzip = 13880
+    VersionClientByteSize     = 36170
+    VersionClientByteSizeGzip = 13899
 
     //region Default durations