+++ site/smoketests/json_merged_signals_test.go
@@ -1,0 +1,9 @@
+package smoketests
+
+import (
+	"testing"
+)
+
+func TestUnitJsonMergedSignals(t *testing.T) {
+	setupPageTestOnLoad(t, "tests/json_merged_signals")
+}