+++ site/smoketests/signal_json_usage_test.go
@@ -1,9 +1,0 @@
-package smoketests
-
-import (
-	"testing"
-)
-
-func TestUnitSignalJsonUsage(t *testing.T) {
-	setupPageTestOnLoad(t, "tests/signal_json_usage")
-}